.ai/pyproject.toml
[project]
name = "orangit-agents"
version = "0.1.0"
description = "OrangIT multi-agent generator"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
orangit-agents-build = "orangit_agents.build:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
